V9-BACKUP-2025-02-14/ai-agents/requirements.txt
# 9LMNTS AI AGENTS - REQUIREMENTS
# Core dependencies for AI automation agents

# HTTP requests and API integration
requests>=2.31.0
urllib3>=1.26.0

# Google Calendar API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0

# Data processing and scheduling
python-dateutil>=2.8.0
pytz>=2023.3

# Environment variable management
python-dotenv>=1.0.0

# JSON handling (built-in)
# No additional packages needed for basic JSON operations

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: For enhanced calendar features
# google-api-core>=2.11.0
# google-auth>=2.22.0
